requirements.txt
tda-api
selenium
dotenv